requirements.txt
python-dotenv==1.0.1
slack_sdk==3.34.0
